Rearrange code as @PoignardAzur suggested
