+++ masonry/src/widget/screenshots/masonry__widget__kurbo__tests__kurbo_shape_circle.png
