+++ xilem/src/view/board/kurbo_shape.rs
@@ -5,41 +5,15 @@
 
 use masonry::widget::{KurboShape, SvgElement};
 use vello::kurbo;
-use xilem_core::{AnyElement, AnyView, DynMessage, MessageResult, Mut, OrphanView, SuperElement};
+use xilem_core::{DynMessage, MessageResult, Mut, OrphanView};
 
 use crate::{Pod, ViewCtx, WidgetView};
-
-use super::BoardElement;
 
 pub trait GraphicsView<State, Action = ()>: WidgetView<State, Action, Widget: SvgElement> {}
 
 impl<V, State, Action> GraphicsView<State, Action> for V where
     V: WidgetView<State, Action, Widget: SvgElement> + Send + Sync
 {
-}
-
-pub type AnyBoardView<State, Action = ()> =
-    dyn AnyView<State, Action, ViewCtx, BoardElement> + Send + Sync;
-
-impl SuperElement<Pod<KurboShape>> for Pod<KurboShape> {
-    fn upcast(child: Pod<KurboShape>) -> Self {
-        child
-    }
-
-    fn with_downcast_val<R>(
-        mut this: Mut<'_, Self>,
-        f: impl FnOnce(Mut<'_, Pod<KurboShape>>) -> R,
-    ) -> (Self::Mut<'_>, R) {
-        let r = f(this.reborrow_mut());
-        (this, r)
-    }
-}
-
-impl AnyElement<Pod<KurboShape>> for Pod<KurboShape> {
-    fn replace_inner(mut this: Self::Mut<'_>, child: Pod<KurboShape>) -> Self::Mut<'_> {
-        this.update_from(child.inner.as_ref().unwrap());
-        this
-    }
 }
 
 macro_rules! impl_orphan_view {
